+++ zetaclient/orchestrator/bootstrap.go
@@ -5,6 +5,7 @@ import (
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/ethclient"
+	ethrpc "github.com/ethereum/go-ethereum/rpc"
 	solrpc "github.com/gagliardetto/solana-go/rpc"
 	"github.com/pkg/errors"
 
@@ -279,12 +280,17 @@ func syncObserverMap(
 				continue
 			}
 
-			// create EVM client
-			evmClient, err := ethclient.DialContext(ctx, cfg.Endpoint)
+			httpClient, err := metrics.GetInstrumentedHTTPClient(cfg.Endpoint)
+			if err != nil {
+				logger.Std.Error().Err(err).Str("rpc.endpoint", cfg.Endpoint).Msgf("Unable to create HTTP client")
+				continue
+			}
+			rpcClient, err := ethrpc.DialHTTPWithClient(cfg.Endpoint, httpClient)
 			if err != nil {
 				logger.Std.Error().Err(err).Str("rpc.endpoint", cfg.Endpoint).Msgf("Unable to dial EVM RPC")
 				continue
 			}
+			evmClient := ethclient.NewClient(rpcClient)
 
 			database, err := db.NewFromSqlite(dbpath, chainName, true)
 			if err != nil {