+++ changelog.md
@@ -5,6 +5,7 @@
 * [2524](https://github.com/zeta-chain/node/pull/2524) - add inscription envolop parsing
 * [2533](https://github.com/zeta-chain/node/pull/2533) - parse memo from both OP_RETURN and inscription
 * [2568](https://github.com/zeta-chain/node/pull/2568) - improve AppContext by converging chains, chainParams, enabledChains, and additionalChains into a single zctx.Chain
+* [2597](https://github.com/zeta-chain/node/pull/2597) - Add generic rpc metrics to zetaclient
 
 
 ## v19.0.1